added themed flash list for auctions
