+++ mobile-app/app/screens/AppNavigator/screens/Auctions/components/VerticalProgressBar.tsx
@@ -1,32 +1,26 @@
 import { View } from "@components";
-import { memo, useState } from "react";
+import { memo } from "react";
 import { useThemeContext } from "@shared-contexts/ThemeProvider";
 import { getColor, tailwind } from "@tailwind";
 import * as Progress from "react-native-progress";
-import { LayoutChangeEvent } from "react-native";
 
 interface VerticalProgressBarProps {
   normalizedBlocks: number;
+  height: number;
   color: string;
 }
 
 export const VerticalProgressBar = memo(
   (props: VerticalProgressBarProps): JSX.Element => {
     const { isLight } = useThemeContext();
-    const [progressBarHeight, setProgressBarHeight] = useState<number>(0);
-
-    const onPageLayout = (event: LayoutChangeEvent) => {
-      const { height } = event.nativeEvent.layout;
-      setProgressBarHeight(height);
-    };
 
     return (
       <View
-        onLayout={onPageLayout}
         style={[
-          tailwind("w-1.5 h-full"),
+          tailwind("w-1.5"),
           {
             transform: [{ rotate: "180deg" }],
+            height: props.height,
           },
         ]}
       >
@@ -34,10 +28,10 @@ export const VerticalProgressBar = memo(
           progress={props.normalizedBlocks}
           style={{
             transform: [
-              { translateX: -(progressBarHeight / 2 - 2) },
+              { translateX: -(props.height / 2 - 2) },
               { translateY: -1 },
               { rotate: "90deg" },
-              { translateX: progressBarHeight / 2 - 2 },
+              { translateX: props.height / 2 - 2 },
               { translateY: -1 },
             ],
           }}
@@ -47,7 +41,7 @@ export const VerticalProgressBar = memo(
           )}
           borderRadius={0}
           borderWidth={0}
-          width={progressBarHeight}
+          width={props.height}
           height={6}
         />
       </View>
